+++ Sources/XcodeMissingFramework/Steps/FileGatherStep.swift
@@ -9,13 +9,12 @@ import Foundation
 import PathKit
 
 public struct FileGatherStep: Step {
-
     public init() {}
-    
+
     public func run(context: StepPipelineContext) throws {
         let files = try context.path.recursiveFilter { $0.isValid(context: context) }
         for file in files {
-            context.files[file.absolute().string] = 0
+            context.files.add(File(filename: file.absolute().string))
         }
     }
 }