+++ Sources/XcodeMissingFramework/Steps/XcodeReferenceGatherStep.swift
@@ -10,7 +10,6 @@ import PathKit
 import xcodeproj
 
 public struct XcodeReferenceGatherStep: Step {
-    
     public init() {}
 
     public func run(context: StepPipelineContext) throws {
@@ -22,15 +21,18 @@ public struct XcodeReferenceGatherStep: Step {
                 }
                 let xcodeProj = try XcodeProj(path: xcodeProjectPath)
                 context.xcodeProjects.append(xcodeProj)
-                let projectPathString = xcodeProjectPath.string.replacingOccurrences(of: xcodeProjectPath.lastComponent, with: "")
+                let last = xcodeProjectPath.lastComponent
+                let projectPathString = xcodeProjectPath.string.replacingOccurrences(of: last, with: "")
                 let projectPath = Path(projectPathString)
-                let xcodeFiles = xcodeProj.pbxproj.fileReferences.compactMap { try? $0.fullPath(sourceRoot: projectPath)?.string }
+                let xcodeFiles = xcodeProj.pbxproj.fileReferences
+                    .compactMap { try? $0.fullPath(sourceRoot: projectPath)?.string }
                 for xcodeFile in xcodeFiles {
                     if context.verbose {
                         print("  Found Xcode File Reference: \(xcodeFile)")
                     }
-                    if let referenceCount = context.files[xcodeFile] {
-                        context.files[xcodeFile] = referenceCount + 1
+                    if var file = context.files.file(for: xcodeFile) {
+                        file.incrementReferenceCount()
+                        context.files.add(file)
                     }
                 }
             } catch {