+++ Sources/xcodemissing/Options/Options.swift
@@ -5,8 +5,8 @@
 //  Created by Jeff Lett on 2/12/19.
 //
 
-import Foundation
 import Commandant
+import Foundation
 import Result
 
 struct Options: OptionsProtocol {
@@ -18,10 +18,14 @@ struct Options: OptionsProtocol {
         return { extensions in { verbose in Options(path: path, extensions: extensions, verbose: verbose) } }
     }
 
-    static func evaluate(_ m: CommandMode) -> Result<Options, CommandantError<CommandantError<()>>> {
+    static func evaluate(_ mode: CommandMode) -> Result<Options, CommandantError<CommandantError<()>>> {
+        let pathDefault = FileManager.default.currentDirectoryPath
+        let pathUsage = "the path to search"
+        let extensionsUsage = "limit the search to files having specific file extensions"
+        let verboseUsage = "enable verbose logging"
         return create
-            <*> m <| Option(key: "path", defaultValue: FileManager.default.currentDirectoryPath, usage: "the path to search")
-            <*> m <| Option(key: "extensions", defaultValue: [], usage: "limit the search to files having specific file extensions")
-            <*> m <| Switch(flag: "v", key: "verbose", usage: "enable verbose logging")
+            <*> mode <| Option(key: "path", defaultValue: pathDefault, usage: pathUsage)
+            <*> mode <| Option(key: "extensions", defaultValue: [], usage: extensionsUsage)
+            <*> mode <| Switch(flag: "v", key: "verbose", usage: verboseUsage)
     }
 }