+++ Tests/LinuxMain.swift
@@ -1,8 +1,0 @@
-import XCTest
-
-import xcodemissingTests
-
-var tests = [XCTestCaseEntry]()
-
-tests += xcodemissingTests.allTests()
-XCTMain(tests)