+++ Tests/XcodeMissingFrameworkTests/Steps/StepPipelineTests.swift
@@ -66,4 +66,19 @@ class StepPipelineTests: XCTestCase {
             XCTAssert(true)
         }
     }
+
+    func testCaseSensitivityDifferenceBetweenXcodeAndFileSystem() {
+        let fileFolder = #file.split(separator: "/").dropLast(3).joined(separator: "/")
+        let path = "/" + fileFolder.appending("/Fixtures/CaseSensitivity/")
+        let steps: [Step] = [FileGatherStep(), XcodeReferenceGatherStep(), UnusedFileAnalyzerStep()]
+        let context = StepPipelineContext(verbose: false, extensions: [".swift"], path: path)
+        let pipeline = StepPipeline(steps: steps, context: context)
+        do {
+            try pipeline.run()
+            print("UnusedFiles: \(context.unusedFiles)")
+            XCTAssert(context.unusedFiles.all.isEmpty)
+        } catch {
+            XCTFail("Should not throw.")
+        }
+    }
 }