+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 #### Bug Fixes
 
-* None.
+* Fixed an issue with false positives caused by Xcode References and the File System having different casing (#9).
 
 ## 0.1.0 (02/15/2019)